% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rvcg-package.R
\docType{data}
\name{dummyhead}
\alias{dummyhead}
\alias{dummyhead.lm}
\alias{dummyhead.mesh}
\title{dummyhead - dummy head and landmarks}
\format{\code{dummyhead.mesh}: triangular mesh representing a dummyhead.

\code{dummyhead.lm}: landmarks on mesh 'dummyhead'}
\description{
A triangular mesh representing a dummyhead - called by data(dummyhead)
}
\keyword{datasets}

