% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{geodesic.path.R}
\alias{geodesic.path.R}
\title{Geodesic path computation with pure R implementation of the backtracking algorithm.}
\usage{
geodesic.path.R(x, source, targets, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d} from the \code{rgl} package.}

\item{source}{scalar positive integer, the source vertex index.}

\item{targets}{positive integer vector, the target vertex indices.}

\item{maxdist}{numeric, the maximal distance to travel along the mesh edges during geodesic distance computation.}
}
\description{
This is an \code{R} implementation of the geodesic path algorithm which is currently used instead of the \code{C++} version since the latter seems to cause a segfault when compiled with the \code{clang} compiler.
}
\keyword{internal}
