% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshInfo.r
\name{meshInfo}
\alias{meshInfo}
\title{print number of vertices and triangular faces of a mesh}
\usage{
meshInfo(x)
}
\arguments{
\item{x}{triangular mesh}
}
\description{
print number of vertices and triangular faces of a mesh
}
