% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshInfo.r
\name{nfaces}
\alias{nfaces}
\title{get number of vertices from a mesh}
\usage{
nfaces(x)
}
\arguments{
\item{x}{triangular mesh}
}
\value{
integer: number of triangular faces
}
\description{
get number of vertices from a mesh
}
