% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshInfo.r
\name{nverts}
\alias{nverts}
\title{get number of vertices from a mesh}
\usage{
nverts(x)
}
\arguments{
\item{x}{triangular mesh}
}
\value{
integer: number of vertices
}
\description{
get number of vertices from a mesh
}
