% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPvdSmooth.r
\name{pvd_smoothnn}
\alias{pvd_smoothnn}
\title{Smooth per vertex data on a mesh.}
\usage{
pvd_smoothnn(mesh, data, num_iter, k = 1L)
}
\arguments{
\item{mesh}{triangular mesh of class \code{mesh3d}}

\item{data}{numeric vector of per vertex data, one scalar value per vertex}

\item{num_iter}{positive integer, the number of iterations}

\item{k}{the k for computing the k-ring-neighborhood. I.e., the max distance (in number of mesh edges) a vertex is allowed to be from the source vertex to be included in its neighborhood.}
}
\value{
numeric vector, the smoothed data
}
\description{
Smooth per vertex data on a mesh.
}
\note{
If you want to smooth several data vectors on the same mesh, it is a lot faster to compute the neighborhood only once using \code{vcgVertexNeighbors} and then use \code{pvd_smoothnn_neigh} many times.
}
\examples{
  data(humface);
  mean_curv = vcgCurve(humface)$meanvb;
  sm_mean_curv = pvd_smoothnn(humface, mean_curv, num_iter=50, k=3);
  hist(mean_curv); hist(sm_mean_curv);
}
