% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPvdSmooth.r
\name{pvd_smoothnn_neigh}
\alias{pvd_smoothnn_neigh}
\title{Perform NN smoothing in given neighborhood.}
\usage{
pvd_smoothnn_neigh(mesh, data, num_iter, neighborhood)
}
\description{
Perform NN smoothing in given neighborhood.
}
