% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvcg-package.R
\docType{package}
\name{Rvcg-package}
\alias{Rvcg}
\alias{Rvcg-package}
\title{Interface between R and vcglib libraries for mesh operations}
\description{
Provides meshing functionality from vcglib (meshlab) for R. E.g. mesh
smoothing, mesh decimation, closest point search.
}
\details{
\tabular{ll}{
Package: \tab Rvcg\cr
Type: \tab Package\cr
Version: \tab 0.24\cr
Date: \tab 2024-09-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\references{
To be announced
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zarquon42b/Rvcg}
  \item \url{https://github.com/cnr-isti-vclab/vcglib}
  \item Report bugs at \url{https://github.com/zarquon42b/Rvcg/issues}
}

}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\keyword{package}
