% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgRaySearch.r
\name{setRays}
\alias{setRays}
\title{helper function to create an object to be processed by vcgRaySearch}
\usage{
setRays(coords, dirs)
}
\arguments{
\item{coords}{k x 3 matrix (or a vector of length 3) containing the starting points of the rays}

\item{dirs}{k x 3 matrix (or a vector of length 3) containing the directons of the rays. The i-th row of \code{dirs} corresponds to the coordinate stored in the i-th row of \code{coords}}
}
\value{
an object of class "mesh3d" (without faces) and the vertices representing the starting points of the rays and the normals storing the directions.
}
\description{
create a search structure from a matrix of coordinates and one of directional vectors to be processed by vcgRaySearch
}
