% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgArea.r
\name{vcgArea}
\alias{vcgArea}
\title{compute surface area of a triangular mesh}
\usage{
vcgArea(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class mesh3d}
}
\value{
surface area of mesh
}
\description{
compute surface area of a triangular mesh
}
\examples{
data(humface)
vcgArea(humface)
}

