% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgCreateKD.r
\name{vcgCreateKDtree}
\alias{vcgCreateKDtree}
\title{create a KD-tree}
\usage{
vcgCreateKDtree(mesh, nofPointsPerCell = 16, maxDepth = 64)
}
\arguments{
\item{mesh}{matrix or triangular mesh containing coordinates}

\item{nofPointsPerCell}{number of points per kd-cell}

\item{maxDepth}{maximum tree depth}
}
\value{
returns an object of class vcgKDtree containing external pointers to the tree and the target points
}
\description{
create a KD-tree
}
\examples{
data(humface)
mytree <- vcgCreateKDtree(humface)
}
\seealso{
\code{\link{vcgSearchKDtree}}
}
