% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgDijkstraPath}
\alias{vcgDijkstraPath}
\title{Compute pseudo-geodesic distance and shortest geodesic path between two vertices on a mesh}
\usage{
vcgDijkstraPath(x, vertpointer_source, vertpointer_target, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{vertpointer_source}{scalar integer, the source vertex index}

\item{vertpointer_target}{scalar integer, the target vertex index}

\item{maxdist}{double, the maximal distance to travel on the mesh}
}
\value{
returns a list of length two, with first entry the path length and second a vector of indices, the path itself.
}
\description{
Compute pseudo-geodesic distance and shortest geodesic path between two vertices on a mesh
}
\note{
Make sure to have a clean manifold mesh. Note that this computes the length of the pseudo-geodesic path (following the edges) between the two vertices.
}
