% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgUpdateNormals.r
\name{vcgFaceNormals}
\alias{vcgFaceNormals}
\title{Compute face normals for a mesh.}
\usage{
vcgFaceNormals(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d', from \code{rgl}}
}
\value{
nx3 matrix of face normals for the mesh
}
\description{
Compute face normals for a mesh.
}
\examples{
data(humface);
hf_facenormals <- vcgFaceNormals(humface);

}
