% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodesicMeanDist}
\alias{vcgGeodesicMeanDist}
\title{Compute, for each mesh vertex, the mean geodesic distance to all other vertices.}
\usage{
vcgGeodesicMeanDist(x, ignore_mask = rep(FALSE, dim(x$vb)[2]))
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{ignore_mask}{logical vector of length \code{|V|}, the number of vertices in the mesh. Each position must indicate whether the vertex should be ignored. Keep in mind that ignoring vertices may lead to a neighborhood consisting of several isolated patches.}
}
\value{
vector of doubles, the mean geodesic distances. The length of the vector equals the number of vertices in the mesh. Ignored vertices will have an \code{NA} value.
}
\description{
Compute, for each mesh vertex, the mean geodesic distance to all other vertices.
}
\examples{
\dontrun{
  fsbrain::download_fsaverage3(TRUE);
  sjd = fsbrain::fsaverage.path();
  sf = subject.surface(sjd, "fsaverage3", "white", "lh");
  mask = fsbrain::subject.mask(sjd, "fsaverage3", hemi = "lh", invert_mask = FALSE);
  tm = fsbrain::fs.surface.to.tmesh3d(sf);
  md = Rvcg::vcgGeodesicMeanDist(tm, ignore_mask = mask);
  fsbrain::vis.data.on.subject(sjd, "fsaverage3", morph_data_lh = md);
}

}
