% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodesicPath}
\alias{vcgGeodesicPath}
\title{Compute geodesic path between vertices on a mesh}
\usage{
vcgGeodesicPath(x, source, targets, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{source}{scalar positive integer, the source vertex index}

\item{targets}{positive integer vector, the target vertex indices}
}
\value{
list of integer vectors, representing the paths.
}
\description{
Compute geodesic path between vertices on a mesh
}
\examples{
data(humface)
vcgGeodesicPath(humface,1,c(5,10))
}
