% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodesicPathA}
\alias{vcgGeodesicPathA}
\title{Compute geodesic path between vertices on a mesh}
\usage{
vcgGeodesicPathA(x, vertpointer_source, vertpointer_targets, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{vertpointer_source}{scalar positive integer, the source vertex index}

\item{vertpointer_targets}{positive integer vector, the target vertex indices}

\item{maxdist}{double, the maximal distance to travel along the mesh edges during distance computation}
}
\value{
an integer vector representing the paths.
}
\description{
Compute geodesic path between vertices on a mesh
}
\examples{
data(humface)
vcgGeodesicPathA(humface,1,c(5,10))
}
