% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgHausdorff.r
\name{vcgHausdorff}
\alias{vcgHausdorff}
\title{calculate the Hausdorff beteween 2 meshes}
\usage{
vcgHausdorff(ref, target, nSamples = 0, nSamplesArea = 0, vertSamp = F,
  edgeSamp = F, faceSamp = F, unrefVert = T, samplingType = c("SS",
  "MC", "SD"), searchStruct = c("SGRID", "AABB", "OCTREE", "HGRID"),
  from = 0, to = 0)
}
\arguments{
\item{ref}{triangular mesh (object of class 'mesh3d').}

\item{target}{triangular mesh (object of class 'mesh3d').}

\item{nSamples}{set the required number of samples.}

\item{nSamplesArea}{set the required number of samples per area unit, override nSamples.}

\item{vertSamp}{logical: if TRUE, disable vertex sampling.}

\item{edgeSamp}{logical: if TRUE, disable edge sampling.}

\item{faceSamp}{logical: if TRUE, disable face sampling.}

\item{unrefVert}{logical: if FLASE, ignore unreferred vertices.}

\item{samplingType}{SS (similar triangles sampling), SD (subdivision sampling), MC (montecarlo sampling).}

\item{searchStruct}{SGIRD (static Uniform Grid), OCTREE, AABB (AxisAligned Bounding Box Tree), HGRID (Hashed Uniform Grid).}

\item{from}{numeric: minimum value for color mapping.}

\item{to}{numeric: maximum value for color mapping.}
}
\value{
\item{ForwardSampling, BackwardSampling}{lists containing information about forward sampling with the following entries}
\itemize{
\item{maxdist}{maximal Hausdorff distance}
\item{meandist}{mean Hausdorff distance}
\item{RMSdist}{RMS of the Hausdorff distances}
\item{nvbsamples}{number of vertices sampled}
\item{nsamples}{number of samples}
}
\item{mesh1, mesh2}{meshes with color coded distances and an additional entry called quality containing the sampled per-vertex distances}
\item{forward_hist, backward_hist}{Matrices tracking the sampling results}
}
\description{
calculate the Hausdorff beteween two triangular meshes.
}

