% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgKmeans.r
\name{vcgKmeans}
\alias{vcgKmeans}
\title{fast Kmean clustering for 2D and 3D data}
\usage{
vcgKmeans(x, k = 10, iter.max = 10, threads = parallel::detectCores())
}
\arguments{
\item{x}{matrix containing coordinates or mesh3d}

\item{k}{number of clusters}

\item{iter.max}{maximum number of iterations}

\item{threads}{integer: number of threads to use}
}
\value{
returns a list containing
\item{centers}{cluster center}
\item{class}{vector with cluster association for each coordinate}
}
\description{
fast Kmean clustering for 2D and 3D data
}
\examples{
require(Rvcg);require(rgl)
data(humface)
set.seed(42)
clust <- vcgKmeans(humface,k=1000,threads=2)
}

