% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPlyWrite.r
\name{vcgObjWrite}
\alias{vcgObjWrite}
\title{Export meshes to OBJ-files}
\usage{
vcgObjWrite(mesh, filename = dataname, writeNormals = TRUE)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d' or a numeric matrix with 3-columns}

\item{filename}{character: filename (file extension '.obj' will be added automatically.}

\item{writeNormals}{write existing normals to file}
}
\description{
Export meshes to OBJ-files
}
\examples{
data(humface)
vcgObjWrite(humface,filename = "humface")
unlink("humface.obj")
}
