% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPlyWrite.r
\name{vcgOffWrite}
\alias{vcgOffWrite}
\title{Export meshes to OFF-files}
\usage{
vcgOffWrite(mesh, filename = dataname)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d' or a numeric matrix with 3-columns}

\item{filename}{character: filename (file extension '.off' will be added automatically.}
}
\description{
Export meshes to OFF-files
}
\examples{
data(humface)
vcgOffWrite(humface,filename = "humface")
unlink("humface.off")
}
