% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgOneRing.r
\name{vcgOneRing}
\alias{vcgOneRing}
\title{Compute one-ring vertex neigbors of vertices.}
\usage{
vcgOneRing(x, vi = NULL, numstep = 1L)
}
\arguments{
\item{x}{tmesh3d instance}

\item{vi}{optional, vector of positive vertex indices for which to compute neighborhood, all are used if left at NULL.}
}
\value{
list of integer vectors, the neighborhoods.
}
\description{
Compute one-ring vertex neigbors of vertices.
}
