% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgSample.r
\name{vcgSample}
\alias{vcgSample}
\title{Subsamples points on a mesh surface}
\usage{
vcgSample(
  mesh,
  SampleNum = 100,
  type = c("km", "pd", "mc"),
  MCsamp = 20,
  geodes = TRUE,
  strict = FALSE,
  iter.max = 100,
  threads = 0
)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d'}

\item{SampleNum}{integer: number of sampled points (see \code{details} below)}

\item{type}{character: select sampling type ("mc"=MonteCarlo Sampling, "pd"=PoissonDisk Sampling,"km"=kmean clustering)}

\item{MCsamp}{integer: MonteCarlo oversampling factor.}

\item{geodes}{logical: maximize geodesic distance between sample points (only for Poisson Disk sampling)}

\item{strict}{logical: if \code{type="pd"} and the amount of coordinates exceeds \code{SampleNum},the resulting coordinates will be subsampled again by kmeans clustering to reach the requested number (Euclidean distances used).}

\item{iter.max}{integer: maximum iterations to use in k-means clustering.}

\item{threads}{integer number of threads to use for k-means clustering}
}
\value{
sampled points
}
\description{
Subsamples surface of a triangular mesh and returns a set of points located on that mesh
}
\details{
The Poisson disk method uses MonteCarlo oversampling followed by Poisson disk pruning. The Poisson disk approach will not generate the exact amount of coordinates specified in \code{SampleNum}, depending on \code{MCsamp} the result will contain more or less coordinates.
}
\examples{

data(humface)
ss <- vcgSample(humface,SampleNum = 500, type="km",threads=1)
\dontrun{
require(rgl)
points3d(ss)
}
}
