% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgSmoothPVD.r
\name{vcgSmoothPVD}
\alias{vcgSmoothPVD}
\title{Smooth per-vertex data on mesh.}
\usage{
vcgSmoothPVD(x, data, fwhm, trunc_factor = 3.5)
}
\arguments{
\item{data}{numeric vector, one value per vertex}

\item{fwhm}{scalar double smoothing kernel full width at half max}

\item{trunc_factor}{scalar double, truncation factor for Gaussian neighborhood}
}
\value{
numeric vector, the smoothed data
}
\description{
Smooth per-vertex data on mesh.
}
