% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPlyWrite.r
\name{vcgStlWrite}
\alias{vcgStlWrite}
\title{Export meshes to STL-files}
\usage{
vcgStlWrite(mesh, filename = dataname, binary = FALSE)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d' or a numeric matrix with 3-columns}

\item{filename}{character: filename (file extension '.stl' will be added automatically.}

\item{binary}{logical: write binary file}
}
\description{
Export meshes to STL-files (binary or ascii)
}
\examples{
data(humface)
vcgStlWrite(humface,filename = "humface")
unlink("humface.stl")
}
