% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vcgUniformRemesh}
\alias{vcgUniformRemesh}
\title{Resample a mesh uniformly}
\usage{
vcgUniformRemesh(x, voxelSize = NULL, offset = 0, discretize = FALSE,
  multiSample = FALSE, absDist = FALSE, mergeClost = FALSE,
  silent = FALSE)
}
\arguments{
\item{x}{triangular mesh}

\item{voxelSize}{voxel size for space discretization}

\item{offset}{see meshlab}

\item{discretize}{see meshlab}

\item{multiSample}{see meshlab}

\item{absDist}{see meshlab}

\item{mergeClost}{logical: merge close vertices}

\item{silent}{logical: suppress messages}
}
\value{
resampled mesh
}
\description{
Resample a mesh uniformly
}

