% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgOneRing.r
\name{vcgVertexNeighbors}
\alias{vcgVertexNeighbors}
\title{Compute mesh adjacency list representation or the vertex neighborhoods of specific mesh vertices.}
\usage{
vcgVertexNeighbors(x, vi = NULL, numstep = 1L, include_self = FALSE)
}
\arguments{
\item{x}{tmesh3d instance}

\item{vi}{optional, vector of positive vertex indices for which to compute neighborhood. All vertices are used if left at the default value \code{NULL}.}

\item{numstep}{positive integer, the number of times to extend the neighborhood (the \code{k} for computing the k-ring neighborhood).}

\item{include_self}{logical, whether the returned neighborhood for a vertex \code{i} should include \code{i} itself.}
}
\value{
list of integer vectors, the neighborhoods.
}
\description{
Compute mesh adjacency list representation or the vertex neighborhoods of specific mesh vertices.
}
\examples{
data(humface)
adjacency_list <- vcgVertexNeighbors(humface)

}
