% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgVolume.r
\name{vcgVolume}
\alias{vcgVolume}
\title{Compute volume for manifold meshes}
\usage{
vcgVolume(x)
}
\arguments{
\item{x}{triangular mesh of class mesh3d}
}
\value{
returns volume
}
\description{
Compute volume for manifold meshes
}
\examples{
mysphere <- vcgSphere()
vcgVolume(mysphere)

}
