% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPlyWrite.r
\name{vcgWrlWrite}
\alias{vcgWrlWrite}
\title{Export meshes to WRL-files}
\usage{
vcgWrlWrite(mesh, filename = dataname, writeCol = TRUE,
  writeNormals = TRUE)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d' or a numeric matrix with 3-columns}

\item{filename}{character: filename (file extension '.wrl' will be added automatically.}

\item{writeCol}{logical: export existing per-vertex color stored in mesh$material$color}

\item{writeNormals}{write existing normals to file}
}
\description{
Export meshes to WRL-files
}
\examples{
data(humface)
vcgWrlWrite(humface,filename = "humface")
unlink("humface.wrl")
}
