\docType{data}
\name{dummyhead}
\alias{dummyhead}
\alias{dummyhead.lm}
\title{dummyhead - dummy head and landmarks}
\format{\code{dummyhead}: triangular mesh representing a dummyhead.

\code{dummyhead.lm}: landmarks on mesh 'dummyhead'}
\description{
A triangular mesh representing a dummyhead - called by
data(dummyhead)
}
\keyword{datasets}

