% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvcg-package.R
\docType{data}
\name{humface}
\alias{humface}
\alias{humface.lm}
\title{Example mesh and landmarks}
\format{\code{humface}: triangular mesh representing a human face.

\code{humface.lm}: landmarks on mesh 'humface'- called by data(humface)}
\description{
A triangular mesh representing a human face - called by data(humface)
}
\keyword{datasets}
