% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rvcg-package.R
\docType{package}
\name{Rvcg-package}
\alias{Rvcg}
\alias{Rvcg-package}
\title{Interface between R and vcglib libraries for mesh operations}
\description{
Provides meshing functionality from vcglib (meshlab) for R. E.g. mesh
smoothing, mesh decimation, closest point search.
}
\details{
\tabular{ll}{
Package: \tab Rvcg\cr
Type: \tab Package\cr
Version: \tab 0.11\cr
Date: \tab 2015-04-14\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

