% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgArea.r
\name{vcgArea}
\alias{vcgArea}
\title{compute surface area of a triangular mesh}
\usage{
vcgArea(mesh, perface = FALSE)
}
\arguments{
\item{mesh}{triangular mesh of class mesh3d}

\item{perface}{logical: if TRUE, a list containing the overall area, as well as the individual per-face area are reported.}
}
\value{
surface area of mesh
}
\description{
compute surface area of a triangular mesh
}
\examples{
data(humface)
vcgArea(humface)
}
