% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgDijkstra}
\alias{vcgDijkstra}
\title{Compute geodesic distances on a triangular mesh}
\usage{
vcgDijkstra(x, vertpointer, tol = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{tol}{numeric: threshold for max distances to consider}

\item{integervector}{references indices of vertices on the mesh}
}
\value{
returns a vector of shortest distances for each of the vertices to one of the vertices referenced in \code{vertpointer}
}
\description{
Compute geodesic distances on a triangular mesh
}
\note{
Make sure to have a clean manifold mesh.
}
\examples{
## Compute geodesic distance between all mesh vertices and the first vertex of a mesh
data(humface)
humface <- vcgIsolated(vcgClean(humface,sel=0:6,iterate=T))
geo <- vcgDijkstra(humface,1)
if (interactive()) {
require(Morpho);require(rgl)
meshDist(humface,distvec = geo)
spheres3d(vert2points(humface)[1,],col=2)
}
}
