% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgDijkstraGeodesicMeanDistPar}
\alias{vcgDijkstraGeodesicMeanDistPar}
\title{Compute geodesic neighborhood in parallel via R parallelization.}
\usage{
vcgDijkstraGeodesicMeanDistPar(x)
}
\description{
Compute geodesic neighborhood in parallel via R parallelization.
}
\examples{
\dontrun{
  fsbrain::download_fsaverage3(TRUE);
  sjd = fsbrain::fsaverage.path();
  sj = "fsaverage3";
  sf = fsbrain::subject.surface(sjd, sj, "white", "lh");
  tm = fsbrain::fs.surface.to.tmesh3d(sf);
  md = vcgDijkstraGeodesicMeanDistPar(tm);
  fsbrain::vis.data.on.subject(sjd, sj, morph_data_lh = md);
}

}
