% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgDijkstraGeodesicMeanDistPar}
\alias{vcgDijkstraGeodesicMeanDistPar}
\title{Compute geodesic neighborhood in parallel via R parallelization.}
\usage{
vcgDijkstraGeodesicMeanDistPar(x, num_cores = NULL)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{num_cores}{scalar positive integer, the number of CPU cores to use. If left at NULL, all cores but one will be used (the default value is \code{max(parallel::detectCores()-1L, 1L)}).}
}
\description{
Compute geodesic neighborhood in parallel via R parallelization.
}
\examples{
\dontrun{
  fsbrain::download_fsaverage3(TRUE);
  sjd = fsbrain::fsaverage.path();
  sj = "fsaverage3";
  sf = fsbrain::subject.surface(sjd, sj, "white", "lh");
  tm = fsbrain::fs.surface.to.tmesh3d(sf);
  md = vcgDijkstraGeodesicMeanDistPar(tm);
  fsbrain::vis.data.on.subject(sjd, sj, morph_data_lh = md);
}

}
