% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgDijkstraPath}
\alias{vcgDijkstraPath}
\title{Compute geodesic path between two vertices on a mesh}
\usage{
vcgDijkstraPath(x, vertpointer_source, vertpointer_target, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{vertpointer_source}{scalar integer, the source vertex index}

\item{vertpointer_target}{scalar integer, the target vertex index}

\item{maxdist}{double, the maximal distance to travel along the mesh edges during distance computation}
}
\value{
an integer vector representing the path.
}
\description{
Compute geodesic path between two vertices on a mesh
}
\examples{
data(humface)
vcgGeodistPath(humface,1,5)
}
