% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodesicMeanDist}
\alias{vcgGeodesicMeanDist}
\title{Compute, for each mesh vertex, the mean geodesic distance to all other vertices.}
\usage{
vcgGeodesicMeanDist(x)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}
}
\value{
vector of doubles, the mean geodesic distances. The length of the vector equals the number of vertices in the mesh.
}
\description{
Compute, for each mesh vertex, the mean geodesic distance to all other vertices.
}
\examples{
\dontrun{
  fsbrain::download_fsaverage3(TRUE);
  sjd = fsbrain::fsaverage.path();
  sf = subject.surface(sjd, "fsaverage3", "white", "lh");
  tm = fsbrain::fs.surface.to.tmesh3d(sf);
  md = Rvcg::vcgGeodesicMeanDist(tm);
  fsbrain::vis.data.on.subject(sjd, "fsaverage3", morph_data_lh = md);
}

}
