% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodesicMeanDist}
\alias{vcgGeodesicMeanDist}
\title{Compute, for each mesh vertex, the mean geodesic distance to all other vertices.}
\usage{
vcgGeodesicMeanDist(x)
}
\arguments{
\item{x}{a tmesh3d instance, see \code{rgl::tmesh3d}.}
}
\value{
vector of doubles, the mean geodesic distances
}
\description{
Compute, for each mesh vertex, the mean geodesic distance to all other vertices.
}
\examples{
\dontrun {
  fsbrain::download_fsaverage3(TRUE);
  sjd = fsbrain::fsaverage.path();
  sf = subject.surface(sjd, "fsaverage3", "white", "lh");
  tm = fsbrain:::ensure.tmesh3d(sf);
  md = Rvcg::vcgGeodesicMeanDist(tm);
  fsbrain::vis.data.on.subject(sjd, "fsaverage3", morph_data_lh = md);
}

}
