% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgDijkstra.r
\name{vcgGeodistPath}
\alias{vcgGeodistPath}
\title{Compute geodesic path between two points on a mesh}
\usage{
vcgGeodistPath(x, pt1, pt2, maxdist = 1e+06)
}
\arguments{
\item{x}{triangular mesh of class \code{mesh3d}}

\item{pt1}{3D coordinate on mesh or index of vertex}

\item{pt2}{3D coordinate on mesh or index of vertex}

\item{maxdist}{double, the maximal distance to travel along the mesh edges during distance computation}
}
\value{
an integer vector representing the path.
}
\description{
Compute geodesic path between two points on a mesh
}
\note{
Make sure to have a clean manifold mesh. Note that this computes the length of the pseudo-geodesic path (following the edges) between the two vertices closest to these points.
}
\examples{
data(humface)
pt1 <- humface.lm[1,]
pt2 <- humface.lm[5,]
vcgGeodistPath(humface,pt1,pt2)
}
