% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcgHausdorff.r
\name{vcgHausdorff}
\alias{vcgHausdorff}
\title{calculate the Hausdorff beteween 2 meshes}
\usage{
vcgHausdorff(ref, target)
}
\arguments{
\item{ref}{triangular mesh (object of class 'mesh3d')}

\item{target}{triangular mesh (object of class 'mesh3d')}
}
\value{
\item{maxdist}{maximal Hausdorff distance}
\item{meandist}{mean Hausdorff distance}
\item{RMSdist}{RMS of the Hausdorff distances}
\item{nvbsamples}{number of vertices sampled}
\item{nsamples}{number of samples}
}
\description{
calculate the Hausdorff beteween two triangular meshes.
}
\examples{
data(humface)
curv <- vcgCurve(humface)
##visualise per vertex mean curvature
\dontrun{
require(Morpho)
meshDist(humface,distvec=curv$meanvb,from=-0.2,to=0.2,tol=0.01)
}
}

