% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgKDtree.r
\name{vcgKDtree}
\alias{vcgKDtree}
\title{perform kdtree search for 3D-coordinates.}
\usage{
vcgKDtree(target, query, k, nofPoints = 16, maxDepth = 64,
  threads = 1)
}
\arguments{
\item{target}{n x 3 matrix with 3D coordinates or mesh of class "mesh3d". These coordinates are to be searched.}

\item{query}{m x 3 matrix with 3D coordinates or mesh of class "mesh3d". We seach the closest coordinates in \code{target} for each of these.}

\item{k}{number of neighbours to find}

\item{nofPoints}{integer: number of points per cell in the kd-tree (don't change unless you know what you are doing!)}

\item{maxDepth}{integer: depth of the kd-tree (don't change unless you know what you are doing!)}

\item{threads}{integer: threads to use in closest point search.}
}
\value{
a list with
\item{index}{integer matrices with indeces of closest points}
\item{distances}{corresponding distances}
}
\description{
perform kdtree search for 3D-coordinates.
}
