% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgOneRing.r
\name{vcgOneRing}
\alias{vcgOneRing}
\title{Compute one-ring vertex neigbors of vertices.}
\usage{
vcgOneRing(x)
}
\arguments{
\item{x}{tmesh3d instance}
}
\value{
list of integer vectors, the neighborhoods. Curently this function still returns FACE indices.
}
\description{
Compute one-ring vertex neigbors of vertices.
}
