% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgPlyWrite.r
\name{vcgPlyWrite}
\alias{vcgPlyWrite}
\alias{vcgPlyWrite.mesh3d}
\alias{vcgPlyWrite.matrix}
\title{Export meshes to PLY-files}
\usage{
vcgPlyWrite(mesh, filename, binary = TRUE, ...)

\method{vcgPlyWrite}{mesh3d}(mesh, filename = dataname, binary = TRUE,
  addNormals = FALSE, writeCol = TRUE, writeNormals = TRUE, ...)

\method{vcgPlyWrite}{matrix}(mesh, filename = dataname, binary = TRUE,
  addNormals = FALSE, ...)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d' or a numeric matrix with 3-columns}

\item{filename}{character: filename (file extension '.ply' will be added automatically, if missing.}

\item{binary}{logical: write binary file}

\item{\dots}{additional arguments, currently not used.}

\item{addNormals}{logical: compute per-vertex normals and add to file}

\item{writeCol}{logical: export existing per-vertex color stored in mesh$material$color}

\item{writeNormals}{write existing normals to file}
}
\description{
Export meshes to PLY-files (binary or ascii)
}
\examples{
data(humface)
vcgPlyWrite(humface,filename = "humface")
## remove it 
unlink("humface.ply")
}
