% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgSample.r
\name{vcgSample}
\alias{vcgSample}
\title{Subsamples points on a mesh surface}
\usage{
vcgSample(mesh, SampleNum = 100, type = c("km", "pd", "mc"), MCsamp = 20,
  geodes = TRUE, strict = FALSE)
}
\arguments{
\item{mesh}{triangular mesh of class 'mesh3d'}

\item{SampleNum}{integer: number of sampled points (see \code{details} below)}

\item{type}{character: seclect sampling type ("mc"=MonteCarlo Sampling, "pd"=PoissonDisk Sampling,"km"=kmean clustering)}

\item{MCsamp}{integer: MonteCarlo sample iterations used in PoissonDisk sampling.}

\item{geodes}{logical: maximise geodesic distance between sample points (only for Poisson Disk sampling)}

\item{strict}{logical: if \code{type="pd"} and the amount of coordinates exceeds \code{SampleNum},  the resulting coordinates will be subsampled again by kmean clustering to reach the requested number.}
}
\value{
sampled points
}
\description{
Subsamples surface of a triangular mesh and returns a set of points located on that mesh
}
\details{
Poisson disk subsampling will not generate the exact amount of coordinates specified in \code{SampleNum}, depending on \code{MCsamp} the result will contain more or less coordinates.
}
\examples{

data(humface)
ss <- vcgSample(humface,SampleNum = 500, type="pd")
\dontrun{
require(rgl)
points3d(ss)
}
}

