\name{vcgVFadj}
\alias{vcgVFadj}
\title{get all faces belonging to each vertex in a mesh}
\usage{
vcgVFadj(mesh)
}
\arguments{
  \item{mesh}{triangular mesh}
}
\value{
list containing one vector per vertex with the indices of
the adjacent faces
}
\description{
get all faces belonging to each vertex in a mesh
}

