% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgVolume.r
\name{vcgVolume}
\alias{vcgVolume}
\title{compute volume of a triangular mesh}
\usage{
vcgVolume(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class mesh3d}
}
\value{
volume of mesh
}
\description{
compute volume of a triangular mesh
}
\examples{

vcgVolume(vcgTetrahedron(normals=TRUE))
}
